+++ response-items/ellipse-scale/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@decsys/ellipse-scale-response",
-  "componentName": "Ellipse Scale",
+  "responseItemName": "Ellipse Scale",
   "version": "2.0.0-alpha.1",
   "description": "The Ellipse Scale Page Response Item for the DECSYS Survey Platform",
   "private": true,
