+++ response-items/ellipse-scale/src/ResponseItem.stats.js
